Change Categorical Values to Numeric

This function converts categorical descriptive values in the columns share, adjustment, and age to numeric values.

Usage

cat_to_num(
  data_participants = read_participants(),
  share_reduced = 0.7,
  adjustment_less = 0.8,
  adjustment_more = 1.2
)

Arguments

data_participants

A dataframe containing the columns share, adjustment, and age.

share_reduced

A numeric value to assign when share is "reduced". Default is 0.7.

adjustment_less

A numeric value to assign when adjustment is "less". Default is 0.8.

adjustment_more

A numeric value to assign when adjustment is "more". Default is 1.2.

Value

A dataframe with transformed values.

Examples

data_participants <- cat_to_num()