Check the dataset for conformity with costsplitter prerequisites

This function checks whether a given dataset conforms to the expected format and requirements of the costsplitter package. It ensures the presence and validity of required columns such as name, pay_, share_, group, age, and adjustment.

Usage

check_costsplitter_data(
  data = readr::read_csv(system.file("participants.csv", package = "costsplitter"),
    show_col_types = FALSE),
  var_pay = "pay_",
  var_share = "share_",
  value_full = 1,
  value_reduced = 0.8,
  value_half = 0.5,
  value_some = 0.3
)

Arguments

data

A data frame containing the cost-sharing data.

var_pay

A string indicating the prefix for payment columns (default is "pay_").

var_share

A string indicating the prefix for share columns (default is "share_").

value_full

A numeric value representing the full share (default is 1).

value_reduced

A numeric value representing the reduced share (default is 0.8).

value_half

A numeric value representing the half share (default is 0.5).

value_some

A numeric value representing the partial share (default is 0.3).

Value

A message indicating whether the dataset conforms to costsplitter prerequisites. Stops execution if the dataset does not conform.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(stringr)
data <- data.frame(
  name = c("Alice", "Bob", "Charlie"),
  pay_breakfast = c(20, 0, 10),
  share_breakfast = c(1, "half", "full"),
  age = c(30, "kid", NA),
  adjustment = c("more", NA, "less")
)

check_costsplitter_data(data)
#>      name pay_breakfast share_breakfast  age adjustment
#> 1   Alice            20               1   30       more
#> 2     Bob             0            half  kid       <NA>
#> 3 Charlie            10            full <NA>       less