Process Age Data in a Data Frame

The helper_process_age function processes the age column in a given data frame. It performs assertive tests to validate the data, converts categorical age values ("adult" and "kid") into numeric equivalents, and scales numeric age values according to specified rules. Additionally, it pivots the data to transform share columns into long format and extracts activities.

Usage

helper_process_age(df = helper_check_names(df))

Arguments

df

A data frame containing an age column and other columns that start with "share". The age column can contain numeric values (0-120) or categorical values ("adult" or "kid").

Value

A tibble with three columns: name, activity, and age. The age values are transformed based on the rules provided, and the data is pivoted to a long format.

Details

The function performs the following steps:

Assertive Tests: Validates that the age column exists and that it contains only numeric values (0-120) or the categorical values "adult" and "kid".

Conversion:

Converts "adult" to 1 and "kid" to 0.5.

Replaces NA and empty values with 1.

Scales numeric values above 1 by dividing by 18.

Data Transformation: Pivots columns that start with "share" into a long format with activity and age columns.

See also

Examples

data("valid_trip_data", package = "costsplitter")
df <- valid_trip_data
data_age <- helper_process_age(df)