Convert Categorical Values to Numeric

This function converts various categorical values within the dataset to numeric values, particularly for cost shares and age adjustments, as per the parameters provided. It processes specific columns like age and adjustment to create numeric equivalents, enabling further analysis and computation within the costsplitter package.

Usage

cat_to_num(
  data = check_costsplitter_data(),
  age_adjustment = (1/18),
  share_adult = 1,
  share_kid = 0.5,
  share_full = 1,
  share_reduced = 0.7,
  share_less = 0.8,
  share_more = 1.2
)

Arguments

data

A data frame, typically the output of check_costsplitter_data(), containing the cost and participant information. Defaults to check_costsplitter_data() if not provided.

age_adjustment

A numeric value used to scale age values to proportions. Default is 1/18.

share_adult

A numeric value indicating the proportion for adults. Default is 1.

share_kid

A numeric value indicating the proportion for kids. Default is 0.5.

share_full

A numeric value representing the full share for participants. Default is 1.

share_reduced

A numeric value representing the reduced share for participants. Default is 0.7.

share_less

A numeric value representing a discounted share applied to participants with the "less" category. Default is 0.8.

share_more

A numeric value representing an increased share applied to participants with the "more" category. Default is 1.2.

Value

A data frame with transformed numeric values for age, adjustment, and other relevant columns.

Details

The function handles several transformations:

Categorical values in cost columns (e.g., "full", "reduced") are converted to numeric.

Age values are translated based on adult and kid patterns in both English and German.

The adjustment column is converted to numeric based on predefined categories like "more" and "less."

Examples

data_participants = check_costsplitter_data() |> cat_to_num()